pyproject.toml
[tool.poetry]
name = "scas"
version = "0.1.0"
description = "Smart Convergence Academic Society"
authors = ["Augustine77 <hugwolf77@gmail.com>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.3"
numpy = "1.26"
pandas = "^2.2.2"
statsmodels = "^0.14.2"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
openpyxl = "^3.1.5"
factor-analyzer = "^0.5.1"
scipy = "^1.14.0"
pyside6 = "^6.7.2"
pyinstaller = "^6.10.0"
psycopg2-binary = "^2.9.9"
ipykernel = "^6.29.5"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
